+++ src/layout/layout-full-window/styles.ts
@@ -4,6 +4,8 @@ export const StyledFullLayout = styled('div')<{ gridTemplate?: string }>`
 	width: 100%;
 	height: 100%;
 
-	display: grid;
-	grid-template-columns: ${(props) => props.gridTemplate ?? '30% 70%;'};
+	display: flex;
+	flex-direction: column;
+
+	gap: 12px;
 `;