+++ src/layout/page-layout/components/content/content.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { PageContentWrapper } from './styles';
+
+export const PageContent: React.FC<{
+	children?: React.ReactElement | React.ReactNode;
+}> = ({ children }) => {
+	return <PageContentWrapper>{children}</PageContentWrapper>;
+};