+++ src/layout/page-layout/components/header/interfaces.ts
@@ -1,0 +1,4 @@
+export interface PageHeaderProps {
+	title?: string | React.ReactElement;
+	children?: React.ReactElement | React.ReactNode;
+}