+++ src/layout/page-layout/components/header/styles.ts
@@ -1,0 +1,7 @@
+import styled from '@emotion/styled';
+
+export const PageHeaderWrapper = styled('div')`
+	width: 100%;
+
+	padding: 12px 12px 12px 0px;
+`;