+++ src/layout/page-layout/index.ts
@@ -1,0 +1,3 @@
+export * from './page-layout';
+export * from './styles';
+export * from './components';