+++ src/layout/page-layout/styles.ts
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+export const PageLayoutWrapper = styled('div')`
+	width: 100%;
+	min-height: 100%;
+
+	display: flex;
+	flex-direction: column;
+`;