+++ src/menu/index.ts
@@ -1,0 +1,4 @@
+export * from './menu';
+export * from './interfaces';
+export * from './styles';
+export * from './components';