+++ src/navigation/components/navigation-footer/index.ts
@@ -1,2 +1,2 @@
-export * from './navigation-option';
 export * from './navigation-footer';
+export * from './styles';