+++ src/navigation/components/navigation-footer/styles.ts
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+
+export const NavigationFooterWrapper = styled('div')`
+	//padding: 12px;
+	margin: 0;
+`;